+++ beecell/logger/event.py
@@ -1,8 +1,8 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
+# SPDX-License-Identifier: EUPL-1.2
 #
-# (C) Copyright 2018-2019 CSI-Piemonte
-# (C) Copyright 2019-2020 CSI-Piemonte
-# (C) Copyright 2020-2021 CSI-Piemonte
+# (C) Copyright 2018-2022 CSI-Piemonte
+
+from beecell.simple import jsonDumps
 
 import time
 import json
@@ -53,7 +53,7 @@ class Event(object):
                'timestamp': self.timestamp, 'proc': self.proc, 'user': self.user,
                'source_addr': self.source_addr, 'dest_addr': self.dest_addr,
                'type': self.type, 'data': self.data}
-        return json.dumps(msg)
+        return jsonDumps(msg)
 
     def publish(self):
         """Publish new event on event logger